requirements.txt
streamlit
youtube_transcript_api
gtts
openai 
requests 
pillow
